feat(api): share membership tables with OSS and fix account-table drift
Moves OrganizationMemberDB/WorkspaceMemberDB/ProjectMemberDB from EE to the
shared OSS models (EE re-exports), creates the three tables in the OSS chain
with the exact EE shapes, backfills membership rows from owners and used
invitations, and aligns api_keys/projects with the shared model in both
chains. get_user_organizations now uses the membership join in both editions.
