+++ api/ee/src/models/db_models.py
@@ -1,118 +1,13 @@
-from datetime import datetime, timezone
+# Membership models moved to OSS (shared schema); re-exported here so existing
+# EE imports keep working.
+from oss.src.models.db_models import (
+    OrganizationMemberDB,
+    WorkspaceMemberDB,
+    ProjectMemberDB,
+)
 
-import uuid_utils.compat as uuid
-
-from sqlalchemy.orm import relationship
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy import Column, String, DateTime, Boolean, ForeignKey
-
-from oss.src.dbs.postgres.shared.base import Base
-
-
-class OrganizationMemberDB(Base):
-    __tablename__ = "organization_members"
-
-    id = Column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid7,
-        unique=True,
-        nullable=False,
-    )
-
-    user_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("users.id", ondelete="CASCADE"),
-    )
-    organization_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("organizations.id", ondelete="CASCADE"),
-    )
-    role = Column(
-        String,
-        nullable=False,
-        server_default="viewer",
-    )
-
-    user = relationship(
-        "oss.src.models.db_models.UserDB",
-    )
-    organization = relationship(
-        "oss.src.models.db_models.OrganizationDB",
-    )
-
-
-class WorkspaceMemberDB(Base):
-    __tablename__ = "workspace_members"
-
-    id = Column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid7,
-        unique=True,
-        nullable=False,
-    )
-
-    user_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("users.id", ondelete="CASCADE"),
-    )
-    workspace_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("workspaces.id", ondelete="CASCADE"),
-    )
-
-    role = Column(String, default="viewer")
-
-    created_at = Column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-    updated_at = Column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-
-    user = relationship(
-        "oss.src.models.db_models.UserDB",
-    )
-    workspace = relationship(
-        "oss.src.models.db_models.WorkspaceDB",
-    )
-
-
-class ProjectMemberDB(Base):
-    __tablename__ = "project_members"
-
-    id = Column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid7,
-        unique=True,
-        nullable=False,
-    )
-
-    user_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("users.id", ondelete="CASCADE"),
-    )
-    project_id = Column(
-        UUID(as_uuid=True),
-        ForeignKey("projects.id", ondelete="CASCADE"),
-    )
-
-    role = Column(String, default="viewer")
-
-    is_demo = Column(Boolean, nullable=True)
-
-    created_at = Column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-    updated_at = Column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-
-    user = relationship(
-        "oss.src.models.db_models.UserDB",
-    )
-    project = relationship(
-        "oss.src.models.db_models.ProjectDB",
-    )
+__all__ = [
+    "OrganizationMemberDB",
+    "WorkspaceMemberDB",
+    "ProjectMemberDB",
+]
