+++ api/oss/src/models/db_models.py
@@ -172,6 +172,115 @@ class ProjectDB(Base):
     )
 
 
+class OrganizationMemberDB(Base):
+    __tablename__ = "organization_members"
+
+    id = Column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid7,
+        unique=True,
+        nullable=False,
+    )
+
+    user_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("users.id", ondelete="CASCADE"),
+    )
+    organization_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("organizations.id", ondelete="CASCADE"),
+    )
+    role = Column(
+        String,
+        nullable=False,
+        server_default="viewer",
+    )
+
+    user = relationship(
+        "oss.src.models.db_models.UserDB",
+    )
+    organization = relationship(
+        "oss.src.models.db_models.OrganizationDB",
+    )
+
+
+class WorkspaceMemberDB(Base):
+    __tablename__ = "workspace_members"
+
+    id = Column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid7,
+        unique=True,
+        nullable=False,
+    )
+
+    user_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("users.id", ondelete="CASCADE"),
+    )
+    workspace_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("workspaces.id", ondelete="CASCADE"),
+    )
+
+    role = Column(String, default="viewer")
+
+    created_at = Column(
+        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
+    )
+    updated_at = Column(
+        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
+    )
+
+    user = relationship(
+        "oss.src.models.db_models.UserDB",
+    )
+    workspace = relationship(
+        "oss.src.models.db_models.WorkspaceDB",
+    )
+
+
+class ProjectMemberDB(Base):
+    __tablename__ = "project_members"
+
+    id = Column(
+        UUID(as_uuid=True),
+        primary_key=True,
+        default=uuid.uuid7,
+        unique=True,
+        nullable=False,
+    )
+
+    user_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("users.id", ondelete="CASCADE"),
+    )
+    project_id = Column(
+        UUID(as_uuid=True),
+        ForeignKey("projects.id", ondelete="CASCADE"),
+    )
+
+    role = Column(String, default="viewer")
+
+    is_demo = Column(Boolean, nullable=True)
+
+    created_at = Column(
+        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
+    )
+    updated_at = Column(
+        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
+    )
+
+    user = relationship(
+        "oss.src.models.db_models.UserDB",
+    )
+    project = relationship(
+        "oss.src.models.db_models.ProjectDB",
+    )
+
+
 class AppDB(Base):
     __tablename__ = "app_db"
